pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ariadne-codegen"
description = "Generate fully typed GraphQL client from schema, queries and mutations!"
authors = [{ name = "Mirumee Software", email = "hello@mirumee.com" }]
version = "0.1.0dev1"
readme = "README.md"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "click~=8.1",
  "graphql-core>=3.2.0,<3.3",
  "toml~=0.10",
  "black~=22.8",
  "isort~=5.10.1",
  "httpx~=0.23",
  "pydantic~=1.10",
  "autoflake~=1.7",
]

[project.optional-dependencies]
dev = [
  "pytest~=7.1.3",
  "pylint~=2.15.3",
  "mypy~=0.971",
  "types-toml~=0.10.8",
  "pytest-mock~=3.8.2",
  "pytest-asyncio~=0.19.0",
  "freezegun~=1.2.2",
]

[project.scripts]
ariadne-codegen = "ariadne_codegen.main:main"

[project.urls]
"Homepage" = "https://github.com/mirumee/ariadne-codegen"
"Bug Tracker" = "https://github.com/mirumee/ariadne-codegen/issues"

[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "missing-docstring",
    "redefined-outer-name",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "too-many-arguments",
    "duplicate-code",
    "no-name-in-module",
    "too-many-locals",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.isort]
profile = "black"

[tool.hatch.build]
include = [
  "ariadne_codegen/**/*.py",
  "ariadne_codegen/py.typed",
  "LICENSE",
  "README.md",
  "EXAMPLE.md",
  "pyproject.toml",
]
exclude = [
  "tests",
]

[ariadne-codegen]
schema_path = "."
queries_path = "."
